Completed videos 36 and 37 - Setting up permission updates on the backend

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const MILLISECS_PER_SEC = 1000;
 const SECS_PER_MIN = 60;
@@ -61,14 +62,21 @@ const Mutations = {
   },
 
   async deleteItem(parent, args, ctx, info) {
+    // throw new Error("You Aren't Allowed to delete this item!");
+
     const where = { id: args.id };
     // 1. Find the item
-    const item = await ctx.db.query.item({ where }, `{id, title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id } }`);
 
     // 2. Check if the request is coming from someone
     // that has permission to delete it
-    //TODO
-
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    if (!(ownsItem || hasPermissions)) {
+      throw new Error("You don't have permission to delete this item.");
+    }
     // 3. Delete the item
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -213,6 +221,41 @@ const Mutations = {
 
     // 8. Return the new user
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in!");
+    }
+
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
